Extract linelog structs to separate header file
